fix: Flow "Cannot resolve module" errors on Windows (#280)
Flow on Windows seems to be unable to correctly resolve imports from `react-css-modules` package, because its `main` entry in `package.json` points to a directory, instead of directly pointing to the script `dist/index.js`. 

Fixing it here seems to be easier than looking for a platform specific issue in Flow.